requirements.txt
pymongo 
python-dotenv  
fastapi
